requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
pymupdf
speechrecognition
matplotlib
